fix: include extra seqs in id_to_name
This commit ensures that if extra spliced and unspliced sequences
are provided, they are added to the id_to_name mapping that is
produced (each extra sequence will map to itself, since there is
no mechanism with extra sequence to assign transcripts to genes).
If one wants a non-trivial sequence-to-gene assignment, the id to
name mapping should be manually modified.